Created a TODO and did some feature design... Mainly just a diary entry
Created in about an hour and a half I mainly rambled my thoughts to help gather my goals for the project. Lots of interesting features can be programmed into this project that may take a long time to implement. One key thought which was figured out at around the end of the document was a keyframe system of lighting effects where you can set a single 'frame' of lighting effects that might take 3 seconds to complete. For example, you make it so every odd numbered led dims down from 100% to 0% in 2 seconds and stays off for 1 second. After this 'frame' of animation is complete than the program runs the second and last 'frame' where the same effect happens for every even numbered led while the odd numbered leds and pulsing back on to whatever you set for them.



With the keyframe system a major question is if each intended lighting sequence should be created with static frames where the timing is adjusted in the blending between the two frames or if each frame should include a single animation sequence for each led and when the program iterates through the next frame there is no option to adjust the timing between the frames and everything is handled in each frame. I hope the two distinct options for keyframe control are apparent here. Having no blending between the keyframes requires more imagination for the user and it would take more skill and probably configuration to make things look right while potentially offering a simpler entry for a high level of control. Having the only 'animation' occur in the settings of blending between two static frames makes everything easier to visualize in the minds eye. However, the first thought in my head is that the blending between two static frames requires much more programming that might be out of scope for this project. The first type of keyframing seems much easier to implement but does not allow the programming to do any "heavy lifting" of the blending for you. My initial thoughts are that I will be going for the first type of keyframing for faster prototyping.